isDownloadRequired is no longer platform-specific
