+++ component/file_cache/file_cache_linux.go
@@ -33,7 +33,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/Seagate/cloudfuse/common"
 	"github.com/Seagate/cloudfuse/common/log"
 	"github.com/Seagate/cloudfuse/internal"
 	"golang.org/x/sys/unix"
@@ -59,72 +58,6 @@ func newObjAttr(path string, info fs.FileInfo) *internal.ObjAttr {
 	}
 
 	return attrs
-}
-
-// isDownloadRequired: Whether or not the file needs to be downloaded to local cache.
-func (fc *FileCache) isDownloadRequired(localPath string, objectPath string, flock *common.LockMapItem) (bool, bool, *internal.ObjAttr, error) {
-	cached := false
-	downloadRequired := false
-	lmt := time.Time{}
-	var stat *syscall.Stat_t = nil
-
-	// check if the file exists locally
-	finfo, statErr := os.Stat(localPath)
-	if statErr == nil {
-		// The file does not need to be downloaded as long as it is in the cache policy
-		fileInPolicyCache := fc.policy.IsCached(localPath)
-		if fileInPolicyCache {
-			cached = true
-		} else {
-			log.Warn("FileCache::isDownloadRequired : %s exists but is not present in local cache policy", localPath)
-		}
-		// gather stat details
-		stat = finfo.Sys().(*syscall.Stat_t)
-		lmt = finfo.ModTime()
-	} else if os.IsNotExist(statErr) {
-		// The file does not exist in the local cache so it needs to be downloaded
-		log.Debug("FileCache::isDownloadRequired : %s not present in local cache", localPath)
-	} else {
-		// Catch all, the file needs to be downloaded
-		log.Debug("FileCache::isDownloadRequired : error calling stat %s [%s]", localPath, statErr.Error())
-	}
-
-	// check if the file is due for a refresh from cloud storage
-	refreshTimerExpired := fc.refreshSec != 0 && time.Since(flock.DownloadTime()).Seconds() > float64(fc.refreshSec)
-
-	// get cloud attributes
-	cloudAttr, err := fc.NextComponent().GetAttr(internal.GetAttrOptions{Name: objectPath})
-	if err != nil && !os.IsNotExist(err) {
-		log.Err("FileCache::isDownloadRequired : Failed to get attr of %s [%s]", objectPath, err.Error())
-	}
-
-	if !cached && cloudAttr != nil {
-		downloadRequired = true
-	}
-
-	if cached && refreshTimerExpired && cloudAttr != nil {
-		// File is not expired, but the user has configured a refresh timer, which has expired.
-		// Does the cloud have a newer copy?
-		cloudHasLatestData := cloudAttr.Mtime.After(lmt) || stat.Size != cloudAttr.Size
-		// Is the local file open?
-		fileIsOpen := flock.Count() > 0 && !flock.LazyOpen
-		if cloudHasLatestData && !fileIsOpen {
-			log.Info("FileCache::isDownloadRequired : File is modified in container, so forcing redownload %s [A-%v : L-%v] [A-%v : L-%v]",
-				objectPath, cloudAttr.Mtime, lmt, cloudAttr.Size, stat.Size)
-			downloadRequired = true
-		} else {
-			// log why we decided not to refresh
-			if !cloudHasLatestData {
-				log.Info("FileCache::isDownloadRequired : File in container is not latest, skip redownload %s [A-%v : L-%v]", objectPath, cloudAttr.Mtime, lmt)
-			} else if fileIsOpen {
-				log.Info("FileCache::isDownloadRequired : Need to re-download %s, but skipping as handle is already open", objectPath)
-			}
-			// As we have decided to continue using old file, we reset the timer to check again after refresh time interval
-			flock.SetDownloadTime()
-		}
-	}
-
-	return downloadRequired, cached, cloudAttr, err
 }
 
 func (fc *FileCache) getAvailableSize() (uint64, error) {
